pyproject.toml
[project]
name = "scrapy-project"
version = "0.1.0"
description = "Fetching data with scrapy"
authors = [
    {name = "MatthieuMordrel",email = "matthieumordrel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "scrapy (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg[binary,pool] (>=3.2.4,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
